+++ src/i_input.cpp
@@ -126,7 +126,7 @@ static int TranslateKey(SDL_Keysym *sym)
         return KEY_RALT;
 
     default:
-        if (scancode >= 0 && scancode < std::size(scancode_translate_table))
+        if (scancode >= 0 && static_cast<size_t>(scancode) < std::size(scancode_translate_table))
         {
             return scancode_translate_table[scancode];
         }
@@ -181,7 +181,7 @@ static int GetTypedChar(SDL_Keysym *sym)
         // If shift is held down, apply the original uppercase
         // translation table used under DOS.
         if ((SDL_GetModState() & KMOD_SHIFT) != 0
-            && result >= 0 && result < std::size(shiftxform))
+            && result >= 0 && static_cast<size_t>(result) < std::size(shiftxform))
         {
             result = shiftxform[result];
         }
